Restructure ipam.go into separate files
